Start transition towards version 3 (broken commit)
There are a bunch of bugs related to the Finnish and English names of
the thesis type, faculty and such now present. These follow from the mix
of new use of keyword arguments using LaTeX3 syntax in tauthesis.cls,
and how the values used to be set using old LaTeX2e methods.

This commit just saves the work in a state that at leat compiles, and
has the keyword arguments present. The above bugs will be fixed in the
following commits on this branch.